kernel/.../{RuleGrammarTest,PriorityAssocTest}: remove PriorityAssocTest, fold into RuleGrammarTest
